backend/requirements.txt
# ==================================================
# ProjectFlow Gantt - Backend Dependencies
# ==================================================

# -------------------------
# Flask Core
# -------------------------
Flask==3.0.0
Werkzeug==3.0.1
Flask-CORS==4.0.0

# -------------------------
# Database
# -------------------------
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7

# -------------------------
# Authentication
# -------------------------
Flask-JWT-Extended==4.6.0
PyJWT==2.8.0

# -------------------------
# Rate Limiting & Redis
# -------------------------
Flask-Limiter==3.5.0
redis==5.0.1

# -------------------------
# Validation & Security
# -------------------------
email-validator==2.1.0
bleach==6.1.0
MarkupSafe==2.1.3

# -------------------------
# Environment
# -------------------------
python-dotenv==1.0.0

# -------------------------
# Production Server
# -------------------------
gunicorn==21.2.0

# -------------------------
# Testing
# -------------------------
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
